[FEATURE] Add zoom/pan interactions and clean native warnings
- Implement pinch-to-zoom and pan gestures in PDF viewer
- Reset zoom state on page/document changes
- Mark roadmap milestone MA.3 complete
- Fix jlong logging format warnings in JNI bridge
- Re-verify debug build success

+++ core/pdf-engine/src/main/cpp/jni_bridge.cpp
@@ -42,7 +42,7 @@ JNIEXPORT jlong JNICALL Java_com_pdfreader_core_pdfengine_MuPdfEngine_nativeOpen
  * Close a PDF document
  */
 JNIEXPORT jint JNICALL Java_com_pdfreader_core_pdfengine_MuPdfEngine_nativeCloseDocument(JNIEnv* env, jobject obj, jlong docHandle) {
-    LOGI("Closing document handle: %lld", docHandle);
+    LOGI("Closing document handle: %ld", static_cast<long>(docHandle));
     
     // TODO: Implement MuPDF document closing
     // fz_drop_document(ctx, doc);
@@ -54,7 +54,7 @@ JNIEXPORT jint JNICALL Java_com_pdfreader_core_pdfengine_MuPdfEngine_nativeClose
  * Get page count
  */
 JNIEXPORT jint JNICALL Java_com_pdfreader_core_pdfengine_MuPdfEngine_nativeGetPageCount(JNIEnv* env, jobject obj, jlong docHandle) {
-    LOGI("Getting page count for document: %lld", docHandle);
+    LOGI("Getting page count for document: %ld", static_cast<long>(docHandle));
     
     // TODO: Implement MuPDF page count retrieval
     // int pages = fz_count_pages(ctx, doc);
